cambios chicos

+++ working/figura.c
@@ -72,13 +72,13 @@ polilinea_t *leer_polilinea(FILE *f){
     return polilinea;
 }
 
-void figura_trasladar(figura_t *figura, double traslacion[2]){
+void figura_trasladar(figura_t *figura, float traslacion[2]){
     for(size_t i=0; i<(figura->cantidad_polilineas);i++){
         polilinea_trasladar(*((figura->polilineas)+i), traslacion);
     }
 }
 
-void figura_rotar(figura_t *figura, double angulo){
+void figura_rotar(figura_t *figura, float angulo){
     for(size_t i=0; i<(figura->cantidad_polilineas);i++){
         polilinea_rotar(*((figura->polilineas)+i), angulo);
     }
